feat(cocoders): upgrade toolchain, align plugins, and refactor TUI/subagent/hooks
Toolchain & Build:
- Upgrade to Rust 1.93.1, update Cargo.lock and workspace dependencies

Skill System:
- SKILL.md frontmatter-based prompt format with metadata parsing
- Skill scanner multi-source discovery and deduplication

Plugin System:
- LSP loader integration for plugin-contributed language servers
- Plugin policy framework and enhanced contributions model
- PluginManifest PLUGIN.toml with skills, hooks, agents, commands, MCP servers

Subagent:
- Plugin-contributed agents with configurable builtins
- Subagent loader for dynamic agent discovery and registration
- Three-layer tool filtering and foreground/background execution

Hook System:
- Context enrichment for pre/post tool-use hooks
- Command handler improvements with scoped priority
- Event consolidation (BeforeToolCall/AfterToolCall/SessionStart)

TUI:
- Shared SuggestionPopup widget extracted for reuse
- Deduplicated popup logic across overlay components

Provider SDKs:
- Code modernization across anthropic, openai, google-genai, volcengine-ark, z-ai
- Aligned provider implementations with updated hyper-sdk traits

Core:
- Compaction improvements (micro-compact, session memory)
- Driver and executor wiring updates
- Tools registry and system-reminder orchestrator refinements

+++ cocode-rs/retrieval/src/reranker/local.rs
@@ -104,7 +104,7 @@ impl LocalReranker {
         // Update scores from reranking results
         for item in &ranked {
             if item.index < results.len() {
-                results[item.index].score = item.score as f32;
+                results[item.index].score = item.score;
             }
         }
